+++ lib/langchain/utils/token_length/base_validator.rb
@@ -21,12 +21,16 @@ module Langchain
 
           leftover_tokens = token_limit(model_name) - text_token_length
 
+          # Some models have a separate token limit for completions (e.g. GPT-4 Turbo)
+          # We want the lower of the two limits
+          max_tokens = [leftover_tokens, completion_token_limit(model_name)].min
+
           # Raise an error even if whole prompt is equal to the model's token limit (leftover_tokens == 0)
-          if leftover_tokens < 0
+          if max_tokens < 0
             raise limit_exceeded_exception(token_limit(model_name), text_token_length)
           end
 
-          leftover_tokens
+          max_tokens
         end
 
         def self.limit_exceeded_exception(limit, length)