+++ spec/langchain/vectorsearch/pgvector_spec.rb
@@ -163,6 +163,7 @@ if ENV["POSTGRES_URL"]
       let(:question) { "How many times is 'lorem' mentioned in this text?" }
       let(:text) { "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed non risus. Suspendisse lectus tortor, dignissim sit amet, adipiscing nec, ultricies sed, dolor." }
       let(:prompt) { "Context:\n#{text}\n---\nQuestion: #{question}\n---\nAnswer:" }
+      let(:response) { double(completion: answer) }
       let(:answer) { "5 times" }
       let(:k) { 4 }
 
@@ -190,11 +191,12 @@ if ENV["POSTGRES_URL"]
 
       context "without block" do
         before do
-          allow(subject.llm).to receive(:chat).with(prompt: prompt).and_return(answer)
+          allow(subject.llm).to receive(:chat).with(prompt: prompt).and_return(response)
+          expect(response).to receive(:context=).with(text)
         end
 
         it "asks a question and returns the answer" do
-          expect(subject.ask(question: question, k: k)).to eq(answer)
+          expect(subject.ask(question: question, k: k).completion).to eq(answer)
         end
       end